PR4: extract TaskRow into pure component + thin authed container (8 actions as callbacks)
